TV: Move SignInViewModel to its own file
